backend/ocr_service/requirements.txt
# Core OCR and PDF Processing
pymupdf>=1.24.0
pdfplumber>=0.11.0
camelot-py[cv]>=0.11.0
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0

# OCR / VLM Clients (Cloud Only)
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.4.0
dashscope>=1.14.0  # For Alibaba Qwen

# Data Processing
pandas>=2.2.0
pydantic>=2.6.0
python-dateutil>=2.8.2

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Supabase Integration
supabase>=2.3.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2
tenacity>=8.2.0  # For retry logic
